pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = ""
authors = ["EliranA <eliranabdoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.32.0"
peft = "^0.5.0"
accelerate = "^0.22.0"
trl = "^0.6.0"
bitsandbytes = "^0.41.1"
pydantic = "^2.3.0"
datasets = "^2.14.4"
torch = "2.0.0"
evaluate = "^0.4.0"
nltk = "^3.8.1"
rogue = "^0.0.2"
rouge-score = "^0.1.2"
scipy= {version="^1.11.2",  python = ">=3.9,<3.13" }
tensorboard = "^2.14.0"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
